create function match and associate new variable matchResult to its output.
function match uses playerSelection and computerSelection variables as
parameters, and has an if ... else structure which takes into account all
possible values from computer and user and its combinations, returning as
output a string that describes what was the matchup and who is the winner
according to it (computer or player).

matchResult is created to store the current value of function match and
is used to display it on the console.