Add do-while loop to execute rounds until a player reaches 3 wins.
The code has been put inside a do-while loop which will only break once
the player or computer scoreUpdated reaches 3.
A few modifications had to be made to make the loop work:
1- variables playerCurrentScore and computerCurrentScore are now created
without value. Their value is given through an if..else statement and
depends on the value of their counterparts playerCurrentScoreUpdated and
computerCurrentScoreUpdated. When the Updated variables have a value
assigned, playerCurrentScore and computerCurrentScore will be equal to
them, otherwise they will have a value of zero.
2- Variables playerCurrentScoreUpdated and computerCurrentScoreUpdated
are now created inside the function score with a value equal to their
playerCurrentScore and computerCurrentScore respectively. This value is
updated within the same function by adding 1 when a player wins a round,
and then is returned.
3- Variables playerCurrentScoreUpdated and computerCurrentScoreUpdated
are now created in the general scope with var instead of let, so their
value can be used retrospectively by their counterparts playerCurrentScore
and computerCurrentScore.

Create and display variable endMatch after the do-while loop breaks.
It will show a message with the final score and declaring the game winner,
depending on which player gets first to three round wins.

Finally, some other corrections where needed to make the program work as
intended:
1- if conditional executable within round function: "console.log" changed
by "return".