Finish style and improve current script
Completed styling according to the projects needs.
Divided "style.css" code for page alignment and organization and for
pure styling.

Improved javascript to accomplish the wanted user experience for the
end-of-the-match-results div, adding 4 new functions ("openPopup",
"hideContent", "removePopup", and "displayContent") which hide or display
the wanted div according to the stage of the game (ongoing or finished).